moved file tests to test_file.py
